+++ batchEdit.ps1
@@ -59,17 +59,16 @@ $editCount = 0
 
 Write-Output "----------------------------------------------" | Tee-Object -file $log_batchEdit
 Write-Output "$(. Get-TimeStamp) CONTENTdm Tools Batch Edit Starting." | Tee-Object -file $log_batchEdit -Append
-Write-Output "Collection Alias: $collection" | Tee-Object -file $log_batchEdit -Append
-Write-Output "Server URL:       $server" | Tee-Object -file $log_batchEdit -Append
-Write-Output "License No.:      $license" | Tee-Object -file $log_batchEdit -Append
-Write-Output "Metadata CSV:     $csv" | Tee-Object -file $log_batchEdit -Append
-Write-Output "CONTENTdm User:   $user" | Tee-Object -file $log_batchEdit -Append
-#THis causes problems when this is run via batch ocr...
-Write-Output ("Executed Command:   " + $MyInvocation.Line) | Tee-Object -FilePath $log_batchEdit -Append
+Write-Output "Collection Alias:`t$collection" | Tee-Object -file $log_batchEdit -Append
+Write-Output "Server URL:`t`t$server" | Tee-Object -file $log_batchEdit -Append
+Write-Output "License No.:`t`t$license" | Tee-Object -file $log_batchEdit -Append
+Write-Output "Metadata CSV:`t`t$csv" | Tee-Object -file $log_batchEdit -Append
+Write-Output "CONTENTdm User:`t`t$user" | Tee-Object -file $log_batchEdit -Append
+Write-Output ("Executed Command:`t"+$MyInvocation.Line) | Tee-Object -FilePath $log_batchEdit -Append
 Write-Output "----------------------------------------------" | Tee-Object -file $log_batchEdit -Append
 
 # Read in the stored org settings to use if no server or license parameters were supplied.
-Write-Debug "Test for existing organizational settings; if they exist, import server and license values."
+Write-verbose "Test for existing organizational settings; if they exist, import server and license values."
 if (Test-Path $cdmt_root\settings\org.csv) {
     $org = $(. Get-Org-Settings)
     if (!($server) -or ($null -eq $server)) {
@@ -80,40 +79,33 @@ if (Test-Path $cdmt_root\settings\org.csv) {
     }
 }
 
+Write-Verbose "Lookup stored user credentials, request if not available."
 # Check for stored user password, if not available get user input.
-Write-Debug "Test for existing user credentials; if they exist use the, if they don't prompt for a password. "
-if (Test-Path $cdmt_root\settings\user.csv) {
-    $usrcsv = $(Resolve-Path $cdmt_root\settings\user.csv)
-    $usrcsv = Import-Csv $usrcsv
-    $usrcsv | Where-Object { $_.user -eq "$user" } | ForEach-Object {
-        $SecurePassword = $_.password | ConvertTo-SecureString
-        $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR($SecurePassword)
-        $pw = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
-        $null = $BSTR
-    }
-    if ("$user" -notin $usrcsv.user) {
-        Write-Output "No user settings found for $user. Enter a password below or store secure credentials using the dashboard." | Tee-Object -Filepath $log_batchEdit -Append
-        [SecureString]$password = Read-Host "Enter $user's CONTENTdm password" -AsSecureString
-        $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR([SecureString]$password)
-        $pw = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
-        $null = $BSTR
-    }
-}
-Else {
-    Write-Output "No user settings file found. Enter a password below or store secure credentials using the dashboard." | Tee-Object -Filepath $log_batchEdit -Append
-    [SecureString]$password = Read-Host "Enter $user's CONTENTdm password" -AsSecureString
-    $BSTR = [System.Runtime.InteropServices.Marshal]::SecureStringToBSTR([SecureString]$password)
-    $pw = [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($BSTR)
-    $null = $BSTR
-}
+$pw = $(. Get-User-Settings -user $user)
 
 # Read in the metadata, add Level column to sort, lookup type, sort rows so Objects are first.
-Write-Debug "Import $csv and read the headers."
+Write-Verbose "Import $csv and read the headers."
 Write-Output "$(. Get-TimeStamp) Reading in the metadata and grouping compound-objects first." | Tee-Object -file $log_batchEdit -Append
 $metadata = Import-Csv -Path "$csv"
 $headers = ($metadata[0].psobject.Properties).Name
 
-Write-Debug "Call the CONTENTdm API for each item to determine its type."
+Write-Verbose "Lookup collection field properties with CONTENTdm API and map field names to nick names, if necessary."
+$props = Invoke-RestMethod "$server/dmwebservices/index.php?q=dmGetCollectionFieldInfo/$collection/json"
+
+foreach ($prop in $props) {
+    if ($prop.name -in $headers) {
+        foreach ($header in $headers) {
+            if ($header -eq $prop.name) {
+                $headers = $headers -replace "$header", "$($prop.nick)"
+            }
+        }
+    }
+}
+
+Write-Verbose "Re-import metadata with the updated headers (nicknames)."
+$metadata = Import-Csv -Path "$csv" -Header $headers | Select-Object -Skip 1
+
+Write-Verbose "Call the CONTENTdm API for each item to determine its type."
 foreach ($record in $metadata) {
     $dmrecord = $record.dmrecord
     $itemInfo = Invoke-RestMethod "$server/dmwebservices/index.php?q=dmGetItemInfo/$collection/$dmrecord/json"
@@ -132,19 +124,19 @@ foreach ($record in $metadata) {
 $metadata2 = New-TemporaryFile
 $metadata | Export-Csv $metadata2 -NoTypeInformation
 
-Write-Debug "Group compound objects first in the metadata to feed to Catcher."
+Write-Verbose "Group compound objects first in the metadata to feed to Catcher."
 $metadata3 = New-TemporaryFile
 $metadata = Import-Csv -Path $metadata2
 $metadata | Sort-Object -Property Level -Descending | Export-Csv $metadata3 -NoTypeInformation
 $metadata = Import-Csv -Path $metadata3
 
-Write-Output "$(. Get-TimeStamp) Processing each metadata update for CONTENTdm Catcher." | Tee-Object -file $log_batchEdit -Append
+Write-Verbose "$(. Get-TimeStamp) Processing each metadata update for CONTENTdm Catcher." | Tee-Object -file $log_batchEdit -Append
 Write-Debug "Establish the loop for building SOAP XML for each update."
 ForEach ($record in $metadata) {
     $dmrecord = $record.dmrecord
     Write-Debug "Building the SOAP for $dmrecord"
     $i++
-    Write-Progress -Activity "Sending Batch Edits to CONTENTdm Catcher" -Status "Sending dmrecord $($collection + "_" + $dmrecord)" -PercentComplete ($i/$metadata.count * 100)
+    Write-Progress -Activity "Sending Batch Edits to CONTENTdm Catcher" -Status "Sending dmrecord $($collection + "_" + $dmrecord)" -PercentComplete ($i / $metadata.count * 100)
     $SOAPRequest = '<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:v6="http://catcherws.cdm.oclc.org/v6.0.0/">' + "`r`n"
     $SOAPRequest += "`t<soapenv:Header/>`r`n"
     $SOAPRequest += "`t<soapenv:Body>`r`n"
@@ -179,9 +171,10 @@ ForEach ($record in $metadata) {
     Try {
         . Send-SOAPRequest -SOAPRequest $SOAPRequest -URL https://worldcat.org/webservices/contentdm/catcher?wsdl -editCount $editCount 2>&1 | Tee-Object -Filepath $log_batchEdit -Append
         $editCount++
-    } Catch {
+    }
+    Catch {
         $editCount--
-        Write-Error "$(. Get-TimeStamp) Error occured, dmrecord $dmrecord not updated." 2>&1  | Tee-Object -Filepath $log_batchEdit -Append
+        Write-Error "$(. Get-TimeStamp) Error occured, dmrecord $dmrecord not updated." 2>&1 | Tee-Object -Filepath $log_batchEdit -Append
         Write-Error $_ 2>&1 | Tee-Object -Filepath $log_batchEdit -Append
         #Write-Host "ERROR ERROR ERROR" -Fore "red" | Tee-Object -file $log_batchEdit -Append
     }